+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
  - New awards (TreeHugger, ThriftShopper, UnderpantsGnome, BavarianGod, ScaredyPants, Thicc)
+ - New stats section "Champions"
 
 
 ## [0.1.0] - 2022-01-15